fix of decision making
There was a bug that it was taking into account all profits from all rounds. Also it was checking only the last won auction. Now there is a loop iterating all won auctions ( a positive value ), plus it checks only auctions won in this round. So it is good finally. Also, i added this option to calculate bidding factor uniquely to each seller and corresponding to them improve the bidding factor. However, i did some tests, but i am not sure 100% if it is definitely correct. Too much time spent.